# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: BehaveDrift API
  version: 1.0.0
  description: |
    **BehaveDrift API** is an open-source AI inference engine that detects
    behavioural pattern drift in dementia residents.

    It accepts longitudinal care observation data and returns:
    - A **drift score** quantifying deviation from each resident's personal baseline
    - **Tiered alerts** (T1–T4) with natural-language explanations
    - **FHIR R4** Condition resources for clinical handover

    ### Authentication

    Two modes are supported:

    **OAuth2 Client Credentials (recommended for production):**
    ```
    POST /v1/auth/token
    grant_type=client_credentials&client_id=...&client_secret=...
    Authorization: Bearer <jwt>
    ```

    **API Key (for self-hosted / development):**
    ```
    X-API-Key: bda_sk_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ### Error Responses
    All errors use a consistent shape:
    ```json
    {
      "error": "validation_error",
      "message": "Field 'resident_id' is required.",
      "code": 422,
      "details": [],
      "request_id": "req_abc123"
    }
    ```

    ### Versioning
    All endpoints are prefixed with `/v1/`. Breaking changes increment to `/v2/`.

  contact:
    name: Umar Isa
    email: hello@behavedrift.io
    url: https://github.com/umarisa/behavedrift-api

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 Client Credentials JWT (RS256, 60 min TTL)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for self-hosted / development use

  # ============================================================
  # SHARED SCHEMAS
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message, code, request_id]
      properties:
        error:
          type: string
          description: Machine-readable error code (snake_case)
          example: validation_error
        message:
          type: string
          description: Human-readable explanation
          example: "Field 'resident_id' is required and cannot be blank."
        code:
          type: integer
          description: HTTP status code
          example: 422
        details:
          type: array
          description: Field-level validation errors (422 only)
          items:
            type: object
            properties:
              field:
                type: string
                example: resident_id
              issue:
                type: string
                example: required
              value:
                nullable: true
                example: null
        request_id:
          type: string
          description: Unique request ID for log correlation
          example: req_4f3a9c21d
        documentation_url:
          type: string
          format: uri
          example: https://docs.behavedrift.io/errors/validation_error

    # ---------- Pagination ----------
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 142
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        has_next:
          type: boolean
          example: true

    # ---------- Resident ----------
...
    AlertSignalDetail:
      type: object
      properties:
        signal:
          type: string
          example: mood
        current_value:
          example: 2
        baseline_mean:
          type: number
          example: 3.7
        baseline_std:
          type: number
          example: 0.6
        z_score:
          type: number
          example: -2.83
        deviation_days:
          type: integer
          example: 3
    AlertExplanation:
      type: object
      properties:
        summary:
          type: string
          example: "Resident is showing a sustained pattern change across 3 signals over the past 72 hours."
        signals:
          type: array
          items:
            $ref: '#/components/schemas/AlertSignalDetail'
        clinical_correlation:
          type: object
          properties:
            pattern:
              type: string
              enum:
                - early_infection_indicator
                - uti_precursor
                - respiratory_precursor
                - psychological_decline
                - positive_improvement
                - medication_side_effect
                - environmental_trigger
                - unknown
              example: early_infection_indicator
            confidence:
              type: number
              minimum: 0
              maximum: 1
              example: 0.63
            suggested_actions:
              type: array
              items:
                type: string
              example:
                - Check fluid intake over past 48 hours
                - Observe for UTI indicators
                - Consider requesting urine dipstick test
                - Notify responsible clinical lead
    AlertResponse:
      type: object
      properties:
        alert_id:
          type: string
          example: alr_3c8e1b90
        resident_id:
          type: string
          example: res_7f3a9c21
        generated_at:
          type: string
          format: date-time
        tier:
          type: string
          enum: [T1, T2, T3, T4]
          example: T2
        tier_label:
          type: string
          enum: [Watch, Concern, Alert, Critical]
          example: Concern
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
          example: 0.74
        drift_score:
          type: number
          minimum: 0
          maximum: 1
          example: 0.67
        explanation:
          $ref: '#/components/schemas/AlertExplanation'
        historical_context:
          type: object
          properties:
            previous_episode_date:
              type: string
              format: date
              nullable: true
              example: "2025-11-03"
            pre_episode_pattern_match_score:
              type: number
              nullable: true
              example: 0.81
        metadata:
          type: object
          properties:
            acknowledged:
              type: boolean
              example: false
            acknowledged_by:
              type: string
              nullable: true
            acknowledged_at:
              type: string
              format: date-time
              nullable: true
            action_taken:
              type: string
              nullable: true
            dismissed:
              type: boolean
              example: false
    AlertAcknowledge:
      type: object
      required: [action_taken]
      properties:
        action_taken:
          type: string
          maxLength: 500
          example: "Requested urine dipstick test. GP notified. Monitoring fluid intake."
        actioned_by:
          type: string
          example: nurse_jane_doe
    AlertDismiss:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 300
          example: "Alert reviewed — resident attended hospital appointment, explaining presentation change."
        dismissed_by:
          type: string
          example: staff_42b

    # ---------- Analytics ----------
    PopulationRiskResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
          example: main_building
        total_residents:
          type: integer
          example: 42
        risk_distribution:
          type: object
          properties:
            stable:     { type: integer, example: 31 }
            watch_t1:   { type: integer, example: 6 }
            concern_t2: { type: integer, example: 4 }
            alert_t3:   { type: integer, example: 1 }
            critical_t4: { type: integer, example: 0 }
        trending_signals:
          type: array
          items:
            type: string
          example: [mood, sleep_quality]
        active_alerts:
          type: array
          items:
            type: object
            properties:
              resident_id: { type: string }
              tier:         { type: string }
              primary_signals:
                type: array
                items:
                  type: string
    ResidentTrendResponse:
      type: object
      properties:
        resident_id:
          type: string
          example: res_7f3a9c21
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        data_points:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              drift_score:
                type: number
              signals:
                type: object
                additionalProperties:
                  type: number
              alert_tier:
                type: string
                nullable: true
                enum: [T1, T2, T3, T4, null]

    # ---------- Webhooks ----------
...
    TokenRequest:
      type: object
      required: [grant_type, client_id, client_secret]
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
        client_id:
          type: string
          example: ten_abc123_client
        client_secret:
          type: string
          example: cs_xxxxxxxxxxxxxxxxxxxx
        scope:
          type: string
          example: "behavedrift:read behavedrift:write"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          example: 3600
        scope:
          type: string
          example: "behavedrift:read behavedrift:write"
    TenantCreate:
      type: object
      required: [organisation_name, contact_email]
      properties:
        organisation_name:
          type: string
          example: Sunrise Care Home
        contact_email:
          type: string
          format: email
          example: admin@sunrisecare.co.uk
        plan:
          type: string
          enum: [self_hosted, starter, enterprise]
          default: self_hosted
    TenantResponse:
      type: object
      properties:
        tenant_id:
          type: string
          example: ten_abc123
        organisation_name:
          type: string
        client_id:
          type: string
          example: ten_abc123_client
        client_secret:
          type: string
          description: Shown only on creation — store securely
          example: cs_xxxxxxxxxxxxxxxxxxxx
        api_key:
          type: string
          description: API Key alternative — shown only on creation
          example: bda_sk_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        message:
          type: string
          example: "Store the client_secret and api_key securely — they will not be shown again."

  # ============================================================
  # SHARED RESPONSE OBJECTS
  # ============================================================
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: authentication_required
            message: A valid Bearer token or X-API-Key header is required.
            code: 401
            request_id: req_abc123
...
    NotFound:
      description: Resource not found (or belongs to another tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: not_found
            message: The requested resource does not exist.
            code: 404
            request_id: req_abc123
    UnprocessableEntity:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: validation_error
            message: "Field 'resident_id' is required."
            code: 422
            details:
              - field: resident_id
                issue: required
                value: null
            request_id: req_abc123
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ============================================================
# PATHS
# ============================================================
paths:
...
  /v1/auth/token:
    post:
      tags: [Auth]
      summary: Issue OAuth2 access token
      description: Exchange client credentials for a short-lived JWT access token (60 min).
      operationId: issueToken
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid_credentials
                message: client_id or client_secret is incorrect.
                code: 401
                request_id: req_abc123
  /v1/auth/tenants:
    post:
      tags: [Auth]
      summary: Provision a new tenant
      description: |
        Creates a new organisation tenant and returns credentials.
        **The client_secret and api_key are shown only once — store them immediately.**

        For self-hosted deployments this endpoint is open.
        For cloud-hosted, requires an admin token.
      operationId: createTenant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '409':
          description: Organisation name already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ---- Residents ----
...
  /v1/alerts:
    get:
      tags: [Alerts]
      summary: List all active alerts
      description: Returns all active (non-dismissed) drift alerts for the tenant.
      operationId: listAlerts
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: tier
          in: query
          description: Filter by minimum alert tier
          schema:
            type: string
            enum: [T1, T2, T3, T4]
        - name: acknowledged
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Active alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/alerts/{resident_id}:
    get:
      tags: [Alerts]
      summary: Get alerts for a resident
      operationId: getResidentAlerts
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: resident_id
          in: path
          required: true
          schema: { type: string }
        - name: include_dismissed
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Resident alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  resident_id: { type: string }
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/alerts/detail/{alert_id}:
    get:
      tags: [Alerts]
      summary: Get a specific alert with full explanation
      operationId: getAlert
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full alert detail with explanation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/alerts/{alert_id}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge an alert
      description: Log that an alert has been reviewed and an action has been taken.
      operationId: acknowledgeAlert
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertAcknowledge'
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Alert already acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/alerts/{alert_id}:
    delete:
      tags: [Alerts]
      summary: Dismiss an alert
      description: Dismiss an alert with a documented reason. Dismissed alerts are retained in audit log.
      operationId: dismissAlert
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertDismiss'
      responses:
        '200':
          description: Alert dismissed
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert_id: { type: string }
                  dismissed: { type: boolean, example: true }
                  dismissed_at: { type: string, format: date-time }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Analytics ----
  /v1/analytics/population:
    get:
      tags: [Analytics]
      summary: Population-level risk distribution
      description: Returns risk distribution and active alerts across all residents for the tenant.
      operationId: getPopulationRisk
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: location
          in: query
          description: Filter by care home location or wing
          schema: { type: string }
        - name: date
          in: query
          description: Snapshot date (defaults to today)
          schema: { type: string, format: date }
      responses:
        '200':
          description: Population risk overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulationRiskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/analytics/trends/{resident_id}:
    get:
      tags: [Analytics]
      summary: Longitudinal drift trend for a resident
      description: Returns time-series drift scores for a resident over the specified period.
      operationId: getResidentTrend
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: resident_id
          in: path
          required: true
          schema: { type: string }
        - name: days
          in: query
          description: Number of days to look back (default 30, max 365)
          schema: { type: integer, default: 30, minimum: 7, maximum: 365 }
      responses:
        '200':
          description: Longitudinal trend data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentTrendResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/analytics/export:
    get:
      tags: [Analytics]
      summary: Export a report
      description: Export observation, alert, or trend data as PDF, CSV, or FHIR Bundle.
      operationId: exportReport
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [pdf, csv, fhir_bundle]
        - name: resident_id
          in: query
          description: Export for a specific resident (omit for full tenant export)
          schema: { type: string }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: include_alerts
          in: query
          schema: { type: boolean, default: true }
        - name: include_observations
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Report file
          content:
            application/pdf:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string }
            application/fhir+json:
              schema:
                type: object
                description: FHIR R4 Bundle
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ---- Webhooks ----
...